refactor(media-library): extract panel drag-drop into a hook
Moves the isDragging state, drag-counter ref, and the four panel drag/drop
handlers (enter/over/leave/drop, including in-app-drag filtering and file-handle
import routing) into use-media-library-drag-drop.ts. Verbatim behaviour.

media-library.tsx: 1771 -> 1694 lines.
